Split API routes into multiple files
